% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_tree_coords.R
\name{export_tree_coords}
\alias{export_tree_coords}
\title{Export Tree Coordinates}
\usage{
export_tree_coords(treetops, chm, filename = "tree_coordinates.csv")
}
\arguments{
\item{treetops}{An sf object with tree tops (from locate_trees).}

\item{chm}{The CHM raster used to extract CRS.}

\item{filename}{Name of the CSV file to save.}
}
\value{
Data frame of exported tree coordinates.
}
\description{
Converts treetop points to lat/lon and exports to CSV.
}
