% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_lidar.R
\name{process_lidar}
\alias{process_lidar}
\title{Process Lidar Data Pipeline with Optional Visualization}
\usage{
process_lidar(las, plot = FALSE)
}
\arguments{
\item{las}{LAS object.}

\item{plot}{Logical. If TRUE, plots intermediate results. Default is FALSE.}
}
\value{
A list with processed elements: DTM, normalized LAS, CHM, treetops, segmented LAS.
}
\description{
Classifies ground, creates DTM, normalizes height, builds and smooths CHM, detects trees, and segments them.
Optionally plots intermediate results.
}
